Fix test in 32 bit python

+++ theano/tests/test_printing.py
@@ -514,8 +514,8 @@ def test_scan_debugprint4():
     def fn(a_m2, a_m1, b_m2, b_m1):
         return a_m1 + a_m2, b_m1 + b_m2
 
-    a0 = theano.shared(numpy.arange(2))
-    b0 = theano.shared(numpy.arange(2))
+    a0 = theano.shared(numpy.arange(2, dtype='int64'))
+    b0 = theano.shared(numpy.arange(2, dtype='int64'))
 
     (a, b), _ = theano.scan(
         fn, outputs_info=[{'initial': a0, 'taps': [-2, -1]},
